fix ioctl_vt: vt_setactivate->console: u32
running `greetd` on big endian fails with errors:
```sh
$ /usr/bin/greetd --config /etc/greetd/config.toml
error: unable to start greeter: terminal: unable to setactivate: ENXIO: No such device or address
terminal: unable to setactivate: ENXIO: No such device or address
error: unable to reset VT: terminal: unable to setactivate: ENXIO: No such device or address
```

This is caused by ioctl.rs using u64
```Rust
pub struct vt_setactivate {
    pub console: u64,
    pub mode: vt_mode,
}
```
while the kernel uses u32 (`linux/vt.h#L83`):
```C
struct vt_setactivate {
	__u32 console;
 	struct vt_mode mode;
};
```
which breaks on big endian due to swapped bytes.

This bug has no effect on little endian, even though the kernel does not
read vt_mode from where it is defined in the struct, but starting 32 bits
earlier at the upper half of the 64bit console variable. But all values
in vt_mode are set to zero and the upper half of vt_mode happens to be
zero too.

Bugs: https://bugs.gentoo.org/979602

+++ greetd/src/terminal/mod.rs
@@ -144,7 +144,7 @@ impl Terminal {
     pub fn vt_setactivate(&self, target_vt: usize) -> Result<(), Error> {
         if cfg!(target_os = "linux") {
             let arg = ioctl::vt_setactivate {
-                console: target_vt as u64,
+                console: target_vt as u32,
                 mode: ioctl::vt_mode {
                     mode: ioctl::VT_AUTO,
                     waitv: 0,
